Changed id-fields

+++ src/main/java/pro/sky/scheduleuniversity/model/Student.java
@@ -7,7 +7,7 @@ public class Student extends Person {
     public Student() {
     }
 
-    public Student(int id, String firstName, String lastName, int age, int groupId) {
+    public Student(Long id, String firstName, String lastName, int age, int groupId) {
         super(id, firstName, lastName, age);
         this.groupId = groupId;
     }
